% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.r
\name{anova_emm_plot}
\alias{anova_emm_plot}
\title{Extract the EMM Plot as a ggplot element form a Jamovi call to ANOVA}
\usage{
anova_emm_plot(res, plotnumber = 1)
}
\arguments{
\item{res}{The result of an ANOVA call}

\item{plotnumber}{The emm plot to extract if using multi-way Anova}
}
\value{
a ggplot object
}
\description{
Extract the EMM Plot as a ggplot element form a Jamovi call to ANOVA
}
